与身份识别算法的通信

+++ BCIMonitor/DataCommunicate/src/tcpcommunicate.h
@@ -13,7 +13,7 @@ public:
     void stop();
     void send(double *data,quint16 data_len);
 signals:
-    void result(quint8);
+    void result(int);
     void readMark(quint8);
 private:
     //数据端口
